Fix shim build

+++ FreeRTOS/Demo/CORTEX_MPS2_QEMU_IAR_GCC/shimlayer.h
@@ -9,7 +9,7 @@
 #define USING_SEMAPHORE   ( 0 )
 #define USING_QUEUE       ( 0 )
 
-size_t shim_send( uint8_t * arr, size_t len );
+int shim_send( uint8_t * arr, size_t len );
 size_t shim_recv(uint8_t * arr, size_t len);
 
 #endif
